Removed unneeded code.
Added __version__ = "0.1.1"

For Python1.5, use importing tools.imputils
if imputil cannot be imported.

Issue a warning instead of raising an exception
if dll's are not found.

Archive is created in bdist_dir, not in collect_dir so
that it is not included into itself.

Print search path of modulefinder.

Distribute the force and debug options also to the install and build
commands.


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk@26 b94a6edd-cd4f-0410-beef-d2780d74e76e