py2exe.py: Now we can build several scripts from one setup() call. The collect-directories are now separate.
The logic behind --includes did not work for dotted module names. Fixed.

support.py:
Added some explaining comments.
Added _finish_import() method to _MyImporter.
imputil seems to have a bug (which I'm not able to find up to now):
Sometimes the line
   if bottom.__ispkg__:
fails with an AttributeError.
This bug can also be triggered by placing the lines:
  import imputil
  imputil._test_revamp()
at the beginning of the script. The bug then also occurrs
on unpacked scripts.

Modules (not packages) imported now ALWAYS get a __file__ attribute (although
this normally does not point to a filename).
Sometimes, in packed programs, we got a traceback ending with the line
'module filename not defined', which was triggered from
Objects/moduleobject.c.


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk/py2exe@70 b94a6edd-cd4f-0410-beef-d2780d74e76e