make_lib_archive no longer appends a .zip extension, this allows other extensions for the archive as well.
Make make_lib_archive a method of the py2exe class, to allow
overriding in a subclass. It also allows the compression to be set -
default is still uncompressed.

The python dll cannot be copied in the same way as other dlls, it must
be in the directory of the executable(s) to be build.


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk@287 b94a6edd-cd4f-0410-beef-d2780d74e76e