The special casing of zlib can be removed.
py2exe arranges that both the shared zipfile *and* the directory it
lives in are on sys.path.  Maybe it would be nicer if only the zipfile
were required on sys.path, but the pythoncom and pywintypes dlls would
not be found then.

The plus is that even compressed shared zipfiles can now be imported
without any special casing of the zlib extension.

Move the 'self.includes.append("warnings")' line into the
plat_prepare() method where it belongs.



Add a comment about a possible improvement.


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk/py2exe@321 b94a6edd-cd4f-0410-beef-d2780d74e76e