run_w.c now creates a _MessageBox Python function and attaches it to the sys module.
boot_common.py uses this, for console-less exes, to display a warning
to the users if there has something been written to stderr (which
ultimately goes into the error logfile) at program exit.  The
_MessageBox function is deleted from the sys module after the boot
code has been executed.

If that's not what the programmer wants, he can of course assign
sys.stderr to something else ;-)


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk/py2exe@334 b94a6edd-cd4f-0410-beef-d2780d74e76e