Fix from Mark Hammond, to correctly manage the thread_state.
He writes: Without this patch, the main thread never releases its
thread state - so as soon as another thread tries to call in, it
deadlocks.  This pretty-much proves that Outlook callbacks are single
threaded :)

I struck this when building a shell extension DLL.  The shell is highly
threaded.  I also added a critical section to the init code, just incase 2
threads try and create the first 2 objects at the same time.


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk/py2exe@412 b94a6edd-cd4f-0410-beef-d2780d74e76e