Monkey patch the builtin compile function to include the filename when it fails.  Problem was that ModuleFinder calls compile and assumes it doesn't fail.  But it fails in strange cases, when files contain NUL bytes, or when a file has an invalid encoding declaration (in the latter case with MemoryError!).  This change at least shows the filename of the module that failed to compile - which may be difficult to find out otherwise.
py2exe patches the version resource string of the python dll in the
dist directory.  The reason for this is that this string is used at
runtime to look up additional paths in the registry, which should of
course not be done for frozen exes.  Since this patching is done every
time py2exe is run, it had also been done if unneeded - and the
patching isn't smart enough to figure out that patching an UPX'd
python dll doesn't really work, and breaks the dll instead.  Now, the
patching is only be done when a fresh copy of pythonxy.dll is copied
into the dist dir.


git-svn-id: https://svn.code.sf.net/p/py2exe/svn/trunk@468 b94a6edd-cd4f-0410-beef-d2780d74e76e